spark-python/pyspark-datasource/pyspark-ds-delta/pyproject.toml
[tool.poetry]
name = "pyspark-ds-delta"
version = "0.1.0"
description = ""
authors = ["mahfooz_iiitian <mahfooz.iiitian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
delta-spark = "^3.1.0"
importlib-metadata = "^7.1.0"
importlib = "^1.0.4"
zipp = "^3.18.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
